refactor: 폴더 구조 변경, 각 기능에 대해 레포지토리 폴더 만듦
