+++ src/main/java/org/example/stockitbe/store/inbound/StoreInboundService.java
@@ -12,14 +12,17 @@ import org.example.stockitbe.store.inbound.model.dto.StoreInboundDto;
 import org.example.stockitbe.store.inbound.model.entity.StoreInboundHeader;
 import org.example.stockitbe.store.inbound.model.entity.StoreInboundItem;
 import org.example.stockitbe.store.inbound.model.entity.StoreInboundStatusHistory;
-import org.example.stockitbe.store.order.StoreOrderHeaderRepository;
-import org.example.stockitbe.store.order.StoreOrderStatusHistoryRepository;
+import org.example.stockitbe.store.inbound.repository.StoreInboundHeaderRepository;
+import org.example.stockitbe.store.inbound.repository.StoreInboundItemRepository;
+import org.example.stockitbe.store.inbound.repository.StoreInboundStatusHistoryRepository;
+import org.example.stockitbe.store.order.repository.StoreOrderHeaderRepository;
+import org.example.stockitbe.store.order.repository.StoreOrderStatusHistoryRepository;
 import org.example.stockitbe.store.order.model.StoreOrderHistoryType;
 import org.example.stockitbe.store.order.model.StoreOrderStatus;
 import org.example.stockitbe.store.order.model.entity.StoreOrderHeader;
 import org.example.stockitbe.store.order.model.entity.StoreOrderStatusHistory;
 import org.example.stockitbe.user.model.AuthUserDetails;
-import org.example.stockitbe.warehouse.outbound.WhOutboundHeaderRepository;
+import org.example.stockitbe.warehouse.outbound.repository.WhOutboundHeaderRepository;
 import org.example.stockitbe.warehouse.outbound.model.entity.WhOutboundHeader;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
