+++ src/main/java/org/example/stockitbe/store/inbound/repository/StoreInboundStatusHistoryRepository.java
@@ -1,4 +1,4 @@
-package org.example.stockitbe.store.inbound;
+package org.example.stockitbe.store.inbound.repository;
 
 import org.example.stockitbe.store.inbound.model.entity.StoreInboundStatusHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
