+++ src/main/java/org/example/stockitbe/store/order/StoreOrderApprovalOrchestrationService.java
@@ -9,18 +9,18 @@ import org.example.stockitbe.hq.infrastructure.model.Infrastructure;
 import org.example.stockitbe.hq.infrastructure.model.StoreWarehouseMap;
 import org.example.stockitbe.hq.infrastructure.model.StoreWarehouseRole;
 import org.example.stockitbe.hq.inventory.InventoryService;
-import org.example.stockitbe.store.inbound.StoreInboundHeaderRepository;
-import org.example.stockitbe.store.inbound.StoreInboundItemRepository;
-import org.example.stockitbe.store.inbound.StoreInboundStatusHistoryRepository;
+import org.example.stockitbe.store.inbound.repository.StoreInboundHeaderRepository;
+import org.example.stockitbe.store.inbound.repository.StoreInboundItemRepository;
+import org.example.stockitbe.store.inbound.repository.StoreInboundStatusHistoryRepository;
 import org.example.stockitbe.store.inbound.model.StoreInboundStatus;
 import org.example.stockitbe.store.inbound.model.entity.StoreInboundHeader;
 import org.example.stockitbe.store.inbound.model.entity.StoreInboundItem;
 import org.example.stockitbe.store.inbound.model.entity.StoreInboundStatusHistory;
 import org.example.stockitbe.store.order.model.entity.StoreOrderHeader;
 import org.example.stockitbe.store.order.model.entity.StoreOrderItem;
-import org.example.stockitbe.warehouse.outbound.WhOutboundHeaderRepository;
-import org.example.stockitbe.warehouse.outbound.WhOutboundItemRepository;
-import org.example.stockitbe.warehouse.outbound.WhOutboundStatusHistoryRepository;
+import org.example.stockitbe.warehouse.outbound.repository.WhOutboundHeaderRepository;
+import org.example.stockitbe.warehouse.outbound.repository.WhOutboundItemRepository;
+import org.example.stockitbe.warehouse.outbound.repository.WhOutboundStatusHistoryRepository;
 import org.example.stockitbe.warehouse.outbound.model.OutboundDestinationType;
 import org.example.stockitbe.warehouse.outbound.model.OutboundSourceType;
 import org.example.stockitbe.warehouse.outbound.model.OutboundStatus;
