+++ src/main/java/org/example/stockitbe/store/order/StoreOrderBatchApproveService.java
@@ -12,6 +12,7 @@ import org.example.stockitbe.store.order.model.StoreOrderBatchTriggerType;
 import org.example.stockitbe.store.order.model.StoreOrderStatus;
 import org.example.stockitbe.store.order.model.dto.StoreOrderBatchDto;
 import org.example.stockitbe.store.order.model.entity.StoreOrderHeader;
+import org.example.stockitbe.store.order.repository.StoreOrderHeaderRepository;
 import org.example.stockitbe.user.model.AuthUserDetails;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
