+++ src/main/java/org/example/stockitbe/store/order/StoreOrderService.java
@@ -22,9 +22,12 @@ import org.example.stockitbe.store.order.model.dto.StoreOrderDto;
 import org.example.stockitbe.store.order.model.entity.StoreOrderHeader;
 import org.example.stockitbe.store.order.model.entity.StoreOrderItem;
 import org.example.stockitbe.store.order.model.entity.StoreOrderStatusHistory;
-import org.example.stockitbe.store.inbound.StoreInboundHeaderRepository;
+import org.example.stockitbe.store.inbound.repository.StoreInboundHeaderRepository;
 import org.example.stockitbe.store.inbound.model.StoreInboundStatus;
 import org.example.stockitbe.store.inbound.model.entity.StoreInboundHeader;
+import org.example.stockitbe.store.order.repository.StoreOrderHeaderRepository;
+import org.example.stockitbe.store.order.repository.StoreOrderItemRepository;
+import org.example.stockitbe.store.order.repository.StoreOrderStatusHistoryRepository;
 import org.example.stockitbe.user.model.AuthUserDetails;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
