+++ src/main/java/org/example/stockitbe/warehouse/outbound/repository/WhOutboundItemRepository.java
@@ -1,4 +1,4 @@
-package org.example.stockitbe.warehouse.outbound;
+package org.example.stockitbe.warehouse.outbound.repository;
 
 import org.example.stockitbe.warehouse.outbound.model.entity.WhOutboundItem;
 import org.springframework.data.jpa.repository.JpaRepository;
