+++ src/main/java/org/example/stockitbe/warehouse/outbound/repository/WhOutboundStatusHistoryRepository.java
@@ -1,4 +1,4 @@
-package org.example.stockitbe.warehouse.outbound;
+package org.example.stockitbe.warehouse.outbound.repository;
 
 import org.example.stockitbe.warehouse.outbound.model.entity.WhOutboundStatusHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
